#161912940 Moved shared validation to middleware - parcelExists and isOwner
