+++ server/src/v1/routes/parcels.js
@@ -1,10 +1,12 @@
 import { Router } from 'express';
 import parcelsController from '../controllers/parcels';
+import Middleware from '../middlewares';
 
 const router = Router();
 router.get('/', parcelsController.index);
-router.get('/:id', parcelsController.getOrder);
-router.put('/:id/cancel', parcelsController.cancelOrder);
-router.put('/:id/destination', parcelsController.changeDestinationOfOrder);
+router.get('/:id', Middleware.parcelExists, parcelsController.getOrder);
+router.put('/:id/cancel', Middleware.parcelExists, Middleware.isOwner, parcelsController.cancelOrder);
+router.put('/:id/destination', Middleware.parcelExists,Middleware.isOwner, parcelsController.changeOrderDestination);
+router.put('/:id/status', Middleware.parcelExists,Middleware.isOwner, parcelsController.changeOrderStatus);
 router.post('/', parcelsController.createOrder);
 export default router;
